throwing an exception if the app doesnt exist to fix a failing test

+++ system-app-servlets/src/main/java/com/caplin/cutlass/app/service/RestApiService.java
@@ -137,9 +137,14 @@ public class RestApiService
 		{
 			destinationWar.delete();
 		}
-		
 		File targetDir = destinationWar.getParentFile();
-		brjs.app(appName).build(targetDir, true);
+		
+		App app = brjs.app(appName);
+		if (!app.dirExists()) {
+			throw new Exception("Unable to export, the app '" + appName + "' doesn't exist.");
+		}
+		
+		app.build(targetDir, true);
 		
 		File tempWar = new File(targetDir, appName + ".war");
 		FileUtils.moveFile(tempWar, destinationWar);
